+++ R/ensure.R
@@ -3,13 +3,14 @@ ensure_not_empty <- function(x) {
 }
 
 ensure_not_dogfood <- function(x) {
-  looks_like_gfm <- any(grepl("^```", x))
-  if (looks_like_gfm) {
+  looks_like_fenced_md <- any(grepl("^```", x))
+  if (looks_like_fenced_md) {
+    lines <- paste0("  ", x[1:3])
     ## I negate yep(), instead of using nope(), to get desired behaviour in
     ## a non-interactive call
     if (!yep(
       "First three lines of putative code are:\n",
-      collapse(newline(x[1:3]), sep = ""),
+      collapse(lines, sep = "\n"), "\n",
       "which doesn't look like R code.\n",
       "Are we going in circles? Did you just run reprex()?\n",
       "In that case, the clipboard now holds the *rendered* result.\n",